requirements.txt
azure-identity
azure-core